python-manager/modules/document-formatter/requirements.txt
requests==2.31.0
python-docx==0.8.11

